Leves alterações legais
adicionei alguns toques finais

+++ client-manager/main.py
@@ -15,22 +15,45 @@ while True:
         name = input("Name: ")
         email = input("Email: ")
         phone = input("Phone: ")
+        
+        if not name or not email:
+            print("⚠️ Nome e email são obrigatórios!")
+            continue
         add_client(name, email, phone)
+        print("✅ cliente adicionado com sucesso!!")
 
     elif choice == "2":
-        for client in list_clients():
-            print(client)
+        clients = list_clients()
+
+        if not clients:
+            print("📭 Nenhum cliente cadastrado.")
+        else:
+            for client in clients:
+                print(f"""
+    ID: {client[0]}
+    Nome: {client[1]}
+    Email: {client[2]}
+    Telefone: {client[3]}
+-------------------------
+""")
 
     elif choice == "3":
-        id = input("ID: ")
+        id = int(input("ID: "))
         name = input("New Name: ")
         email = input("New Email: ")
-        phone = input("New Phone: ") 
+        phone = input("New Phone: ")
+
         update_client(id, name, email, phone)
+        print("✏️ Cliente atualizado com sucesso!!")
 
     elif choice == "4":
-        id = input("ID: ")
-        delete_client(id)
+        id = int(input("ID: "))
+
+        confirm = input("Tem certeza que deseja deletar? (s/n):")
+        
+        if confirm.lower() == "s":
+            delete_client(id)
+            print("🗑️ Cliente removido com sucesso!!")
 
     elif choice == "5":
         break
